Split to create a library
